refactor: split safe access traits into microcrate
