+++ crates/copybook-safe-access/Cargo.toml
@@ -1,0 +1,27 @@
+# SPDX-License-Identifier: AGPL-3.0-or-later
+[package]
+name = "copybook-safe-access"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Panic-safe Option/Vec/slice access traits for copybook-rs"
+documentation = "https://docs.rs/copybook-safe-access"
+readme = "README.md"
+keywords = ["cobol", "copybook", "safety", "error-handling"]
+categories = ["rust-patterns", "data-structures"]
+
+include = ["src/**", "tests/**", "Cargo.toml", "README.md", "LICENSE"]
+
+[package.metadata.docs.rs]
+all-features = false
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+copybook-error.workspace = true
+
+[lints]
+workspace = true
