+++ crates/copybook-safe-access/README.md
@@ -1,0 +1,3 @@
+# copybook-safe-access
+
+Panic-safe extension traits for `Option`, `Vec`, and slices that return `copybook-error` values with contextual messages.
